refactor: 로딩스피너 대신 결과 추첨 화면으로 fallback UI 대체
