+++ frontend/src/pages/pickeat/matchResult/MatchResult.tsx
@@ -1,11 +1,24 @@
+import ResultRestaurantCard from '@domains/pickeat/matchResult/components/ResultRestaurantCard';
+
 import { HEADER_HEIGHT } from '@widgets/Header';
+
+import RoundedButton from '@components/actions/RoundedButton';
+import Confetti from '@components/Confetti';
+
+import { ApiError } from '@apis/apiClient';
+import { pickeat } from '@apis/pickeat';
+
+import { ROUTE_PATH } from '@routes/routePath';
+
+import { useShowToast } from '@provider/ToastProvider';
 
 import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
-import { useState } from 'react';
+import { Suspense, useCallback, useMemo } from 'react';
+import { useNavigate, useSearchParams } from 'react-router';
 
 import PendingResultScreen from './components/PendingResultScreen';
-import ResultScreen from './components/ResultScreen';
+import Twinkles from './components/Twinkles';
 
 const scaleUp = keyframes`
   from {
@@ -19,16 +32,55 @@ const scaleUp = keyframes`
 `;
 
 function MatchResult() {
-  const [readied, setReadied] = useState(false);
+  const navigate = useNavigate();
+  const [searchParams] = useSearchParams();
+  const pickeatCode = searchParams.get('code') ?? '';
+  const showToast = useShowToast();
+
+  const fetchResult = useCallback(async () => {
+    try {
+      const actualResultPromise = pickeat.getResult(pickeatCode);
+      const delayPromise = new Promise(resolve => setTimeout(resolve, 2500));
+      const [result] = await Promise.all([actualResultPromise, delayPromise]);
+
+      if (!result) throw new Error('투표 결과가 없습니다.');
+
+      return result;
+    } catch (e) {
+      if (e instanceof ApiError && e.status === 401) {
+        showToast({
+          mode: 'ERROR',
+          message: '해당 픽잇에 접근할 수 없습니다.',
+        });
+        navigate(ROUTE_PATH.MAIN);
+      } else {
+        showToast({
+          mode: 'ERROR',
+          message: '투표 결과를 불러오지 못했습니다!',
+        });
+      }
+      throw e;
+    }
+  }, []);
+
+  const restaurantData = useMemo(() => fetchResult(), []);
+
   return (
     <S.Container>
-      {readied ? (
+      <Suspense fallback={<PendingResultScreen />}>
         <S.ResultScreenWrapper>
-          <ResultScreen />
+          <S.ResultWrapper>
+            <Confetti />
+            <ResultRestaurantCard restaurantData={restaurantData} />
+            <S.TwinkleBox>
+              <Twinkles />
+            </S.TwinkleBox>
+          </S.ResultWrapper>
+          <RoundedButton onClick={() => navigate(ROUTE_PATH.MAIN)}>
+            메인으로 돌아가기
+          </RoundedButton>
         </S.ResultScreenWrapper>
-      ) : (
-        <PendingResultScreen onReady={() => setReadied(true)} />
-      )}
+      </Suspense>
     </S.Container>
   );
 }
@@ -43,11 +95,42 @@ const S = {
     justify-content: center;
     align-items: center;
 
-    padding-top: ${HEADER_HEIGHT};
-
     background-color: ${({ theme }) => theme.PALETTE.gray[5]};
   `,
   ResultScreenWrapper: styled.div`
+    height: 100vh;
+    padding-top: ${HEADER_HEIGHT};
+
     animation: ${scaleUp} 0.3s ease forwards;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: ${({ theme }) => theme.GAP.level8};
+  `,
+  ResultWrapper: styled.div`
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: ${({ theme }) => theme.GAP.level2};
+    position: relative;
+
+    border-radius: ${({ theme }) => theme.RADIUS.xlarge};
+    box-shadow: none;
+  `,
+  TwinkleBox: styled.div`
+    width: 270px;
+    height: 300px;
+    position: absolute;
+    pointer-events: none;
+  `,
+  Title: styled.p`
+    color: ${({ theme }) => theme.PALETTE.gray[60]};
+    font: ${({ theme }) => theme.FONTS.heading.large};
+  `,
+  ToMainButton: styled.div`
+    width: 260px;
   `,
 };
